# swagger.yaml
openapi: 3.0.0
components:
...
  schemas:
...
    _36_Enums.TravelStatus:
      type: string
      enum:
        - IN_PROGRESS
        - COMPLETED
        - FAILED
    TravelStatus:
      $ref: "#/components/schemas/_36_Enums.TravelStatus"
    TravelResultSummary:
      properties:
        id:
          anyOf:
            - type: string
            - type: number
              format: double
        userId:
          type: number
          format: double
        title:
          type: string
        summaryContent:
          type: string
        recommendations:
          items:
            type: string
          type: array
      required:
        - id
        - userId
        - title
        - summaryContent
        - recommendations
      type: object
      additionalProperties: false
    PlanetTravelStartApiResponse:
      properties:
        travelId:
          anyOf:
            - type: string
            - type: number
              format: double
        travelResultId:
          anyOf:
            - type: string
            - type: number
              format: double
        remainingFuel:
          type: number
          format: double
        status:
          anyOf:
            - $ref: "#/components/schemas/TravelStatus"
            - type: string
        travelResult:
          $ref: "#/components/schemas/TravelResultSummary"
      required:
        - travelId
        - travelResultId
        - remainingFuel
        - status
      type: object
      additionalProperties: false
    ApiResponse_PlanetTravelStartApiResponse_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/PlanetTravelStartApiResponse"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    _36_Enums.PlanetType:
      type: string
      enum:
        - MEAL
        - WATER
        - EMOTION
        - LIFESTYLE
        - RETROSPECT
    PlanetType:
      $ref: "#/components/schemas/_36_Enums.PlanetType"
    PlanetTravelStartApiRequest:
      properties:
        planetType:
          anyOf:
            - $ref: "#/components/schemas/PlanetType"
            - type: string
              enum:
                - MEAL
                - WATER
                - EMOTION
                - LIFESTYLE
                - RETROSPECT
        fuelSpent:
          type: number
          format: double
      required:
        - planetType
        - fuelSpent
      type: object
      additionalProperties: false
    TravelStateInfoResponse:
      properties:
        currentPlanet:
          type: string
        explorationProgressPercent:
          type: number
          format: double
        currentFuel:
          type: number
          format: double
        requiredFuelForNextPlanet:
          type: number
          format: double
        tammyRelationshipLevel:
          type: number
          format: double
      required:
        - explorationProgressPercent
        - currentFuel
        - requiredFuelForNextPlanet
        - tammyRelationshipLevel
      type: object
      additionalProperties: false
    ApiResponse_TravelStateInfoResponse_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/TravelStateInfoResponse"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    TravelResultDetailInfo:
      properties:
        reportId:
          anyOf:
            - type: string
            - type: number
              format: double
        travelResultId:
          anyOf:
            - type: string
            - type: number
              format: double
        userId:
          type: number
          format: double
        title:
          type: string
        summaryContent:
          type: string
        recommendations:
          items:
            type: string
          type: array
        createdAt:
          anyOf:
            - type: string
            - type: string
              format: date-time
      required:
        - title
        - summaryContent
        - recommendations
      type: object
      additionalProperties: false
    ApiResponse_TravelResultDetailInfo_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/TravelResultDetailInfo"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    CalorieTrendItem:
      properties:
        date:
          type: string
        caloriesKcal:
          type: number
          format: double
      required:
        - date
        - caloriesKcal
      type: object
      additionalProperties: false
    NutritionBalanceInfo:
      properties:
        carbohydratePercent:
          type: number
          format: double
        proteinPercent:
          type: number
          format: double
        fatPercent:
          type: number
          format: double
        vitaminPercent:
          type: number
          format: double
        mineralPercent:
          type: number
          format: double
      required:
        - carbohydratePercent
        - proteinPercent
        - fatPercent
        - vitaminPercent
        - mineralPercent
      type: object
      additionalProperties: false
    DashboardSummaryInfo:
      properties:
        calorieTrends:
          items:
            $ref: "#/components/schemas/CalorieTrendItem"
          type: array
        nutritionBalance:
          $ref: "#/components/schemas/NutritionBalanceInfo"
        weeklyWorkoutCompletedDays:
          type: number
          format: double
      required:
        - calorieTrends
        - nutritionBalance
        - weeklyWorkoutCompletedDays
      type: object
      additionalProperties: false
    ApiResponse_DashboardSummaryInfo_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/DashboardSummaryInfo"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
...
    ApiResponse_any_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
...
    ChatMessageApiRequest:
      properties:
        messageText:
          type: string
      required:
        - messageText
      type: object
      additionalProperties: false
    ApiResponse_null_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          anyOf: []
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    _36_Enums.AuthProvider:
      type: string
      enum:
        - LOCAL
        - KAKAO
        - GOOGLE
        - APPLE
    AuthProvider:
      $ref: "#/components/schemas/_36_Enums.AuthProvider"
    UserAuthProfile:
      properties:
        id:
          type: number
          format: double
        email:
          type: string
        nickname:
          type: string
        authProvider:
          anyOf:
            - $ref: "#/components/schemas/AuthProvider"
            - type: string
              enum:
                - LOCAL
                - KAKAO
                - GOOGLE
                - APPLE
      required:
        - id
        - email
        - nickname
        - authProvider
      type: object
      additionalProperties: false
    UserLoginResponse:
      properties:
        user:
          $ref: "#/components/schemas/UserAuthProfile"
        accessToken:
          type: string
          description: 인증 성공 후 발급되는 JWT Access Token
          example: eyJhbGciOiJIUzI1Ni...
        refreshToken:
          type: string
          description: 토큰 갱신용 JWT Refresh Token
          example: eyJhbGciOiJIUzI1Ni...
      required:
        - user
        - accessToken
        - refreshToken
      type: object
      additionalProperties: false
    ApiResponse_UserLoginResponse_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/UserLoginResponse"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    UserSignUpRequest:
      properties:
        email:
          type: string
          description: 가입 및 로그인에 사용될 이메일 주소입니다.
          example: user@example.com
          pattern: ^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$
        password:
          type: string
          description: 보안성이 뛰어난 비밀번호입니다. (최소 8자, 최대 30자)
          example: Password123!
          minLength: 8
          maxLength: 30
        nickname:
          type: string
          description: 사용자 닉네임입니다. (최소 2자, 최대 20자)
          example: 우주탐험가
          minLength: 2
          maxLength: 20
        gender:
          type: string
          description: 성별 (MALE, FEMALE, UNKNOWN)
        age:
          type: number
          format: double
          description: 나이
      required:
        - email
        - password
        - nickname
      type: object
      additionalProperties: false
    UserLoginRequest:
      properties:
        email:
          type: string
          description: 로그인용 이메일 주소
          example: user@example.com
        password:
          type: string
          description: 로그인용 비밀번호
          example: Password123!
      required:
        - email
        - password
      type: object
      additionalProperties: false
    SocialLoginRequest:
      properties:
        provider:
          type: string
          enum:
            - GOOGLE
            - KAKAO
            - APPLE
          description: 소셜 로그인 제공자 (GOOGLE | KAKAO | APPLE)
          example: KAKAO
        token:
          type: string
          description: 구글/카카오/애플 소셜 로그인 인증 토큰 (access_token 또는 id_token)
          example: ya29.a0AfH6SM...
        nickname:
          type: string
          description: 신규 가입 시 기본 설정할 닉네임 (선택)
          example: 우주탐험가
      required:
        - provider
        - token
      type: object
      additionalProperties: false
    TokenRefreshResponse:
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - accessToken
        - refreshToken
      type: object
      additionalProperties: false
    ApiResponse_TokenRefreshResponse_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/TokenRefreshResponse"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    TokenRefreshRequest:
      properties:
        refreshToken:
          type: string
          description: Access Token 재발급용 Refresh Token
          example: eyJhbGciOiJIUzI1Ni...
      required:
        - refreshToken
      type: object
      additionalProperties: false
    UserWithdrawResponse:
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
      type: object
      additionalProperties: false
    ApiResponse_UserWithdrawResponse_:
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          allOf:
            - $ref: "#/components/schemas/UserWithdrawResponse"
          nullable: true
        code:
          type: number
          format: double
      required:
        - success
        - message
        - data
        - code
      type: object
      additionalProperties: false
    UserWithdrawRequest:
      properties:
        reason:
          type: string
          description: 탈퇴 사유
          example: 개인 사정
      type: object
      additionalProperties: false
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
info:
  title: nasa_backend
  version: 1.0.0
  description: NASA Wellness Backend API System with Tammy AI Virtual Pet
  license:
    name: ISC
  contact: {}
paths:
...
  /planet-travel/start:
    post:
      operationId: StartPlanetTravel
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_PlanetTravelStartApiResponse_"
      description: 보유한 우주 연료(Fuel)를 소모하여 선택한 행성으로 별여행을 출발하고, AI 탐사 결과(travelResult)
        진단서를 실시간 동기 생성합니다.
      summary: 별여행 탐사 출발 및 AI 탐사결과 생성
      tags:
        - 6. PlanetTravel - 별여행 게이미피케이션 탐사 & 진단서
      security:
        - jwt: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanetTravelStartApiRequest"
  /planet-travel/state:
    get:
      operationId: GetTravelState
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_TravelStateInfoResponse_"
      description: 우주여행 현황 조회
      summary: 우주여행 현황 조회
      tags:
        - 6. PlanetTravel - 별여행 게이미피케이션 탐사 & 진단서
      security:
        - jwt: []
      parameters: []
  /travel-results/{travelResultId}:
    get:
      operationId: GetTravelResult
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_TravelResultDetailInfo_"
      description: 행성 탐사 완료 후 생성된 AI 별여행 탐사 결과 진단서(travelResult)의 상세 보고서 내용을 조회합니다.
      summary: AI 별여행 탐사 결과 진단서 상세 조회
      tags:
        - 6. PlanetTravel - 별여행 게이미피케이션 탐사 & 진단서
      security:
        - jwt: []
      parameters:
        - in: path
          name: travelResultId
          required: true
          schema:
            type: string
  /reports/{reportId}:
    get:
      operationId: GetReport
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_TravelResultDetailInfo_"
      description: AI 별여행 탐사 결과 진단서 상세 조회 (호환용 라우트)
      summary: 보고서 상세 조회 (호환용)
      tags:
        - 6. PlanetTravel - 별여행 게이미피케이션 탐사 & 진단서
      security:
        - jwt: []
      parameters:
        - in: path
          name: reportId
          required: true
          schema:
            type: string
  /dashboard/summary:
    get:
      operationId: GetDashboardSummary
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_DashboardSummaryInfo_"
      description: 주간 칼로리 섭취 추이 및 3대 영양소 밸런스 비율 통계 대시보드를 조회합니다.
      summary: 웰니스 종합 대시보드 통계 조회
      tags:
        - 6. PlanetTravel - 별여행 게이미피케이션 탐사 & 진단서
      security:
        - jwt: []
      parameters: []
...
  /chat/message:
    post:
      operationId: SendMessage
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_any_"
      description: AI 버추얼 펫 타미(TAMMY)와 실시간 심리 공감 대화를 나누고, 모션 태그 응답 수령 및 장기 기억 캡슐을 생성합니다.
      summary: AI 타미 심리 공감 메시지 전송
      tags:
        - 5. TammyChat - AI 타미 심리 공감 대화
      security:
        - jwt: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageApiRequest"
  /chat/messages/{messageId}:
    delete:
      operationId: DeleteMessage
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_null_"
      description: "[3.4] 대화 메시지 소프트 삭제 API"
      tags:
        - 5. TammyChat - AI 타미 심리 공감 대화
      security:
        - jwt: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
  /chat/messages/{messageId}/undo:
    post:
      operationId: UndoDeleteMessage
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_null_"
      description: "[3.4] 대화 메시지 삭제 취소 (Undo) API"
      tags:
        - 5. TammyChat - AI 타미 심리 공감 대화
      security:
        - jwt: []
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: string
  /auth/register:
    post:
      operationId: Register
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserLoginResponse_"
      description: 이메일과 비밀번호 기반으로 신규 회원을 등록합니다.
      summary: 일반 회원가입
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUpRequest"
  /auth/signup:
    post:
      operationId: SignUp
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserLoginResponse_"
      description: "[3.6] 회원가입 API (호환용)"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUpRequest"
  /auth/login:
    post:
      operationId: Login
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserLoginResponse_"
      description: "[3.6] 로그인 API"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
  /auth/social-login:
    post:
      operationId: SocialLogin
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserLoginResponse_"
      description: "[3.6] 구글 / 카카오 / 애플 소셜 로그인 API"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialLoginRequest"
  /auth/refresh:
    post:
      operationId: Refresh
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_TokenRefreshResponse_"
      description: "[3.6] 토큰 자동 갱신 API"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
  /auth/withdraw:
    post:
      operationId: WithdrawPost
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserWithdrawResponse_"
      description: "[3.6] 회원 탈퇴 API"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security:
        - jwt: []
      parameters: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWithdrawRequest"
    delete:
      operationId: Withdraw
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_UserWithdrawResponse_"
      tags:
        - 1. Auth - 회원 인증 및 소셜 로그인
      security:
        - jwt: []
      parameters: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserWithdrawRequest"
servers:
  - url: /api